Moved task_map headers
