+++ exotations/task_maps/task_map/src/EffOrientation.cpp
@@ -30,7 +30,7 @@
  *
  */
 
-#include "EffOrientation.h"
+#include "task_map/EffOrientation.h"
 
 REGISTER_TASKMAP_TYPE("EffOrientation", exotica::EffOrientation);
