+++ exotations/task_maps/task_map/src/IMesh.cpp
@@ -30,7 +30,7 @@
  *
  */
 
-#include "IMesh.h"
+#include "task_map/IMesh.h"
 #define XML_CHECK(x) {xmltmp=handle.FirstChildElement(x).ToElement();if (!xmltmp) throw_named("XML element '"<<x<<"' does not exist!");}
 
 //#define DEBUG_MODE
